+++ tests/agent_tests.py
@@ -6,6 +6,7 @@ import uuid
 
 import consulate
 from consulate import utils
+from consulate.models import agent
 
 from . import base
 
@@ -121,32 +122,32 @@ class CheckTestCase(base.TestCase):
 
     def test_register(self):
         self.assertTrue(self.consul.agent.check.register(
-            str(uuid.uuid4()), http='http://localhost', interval=30))
+            str(uuid.uuid4()), http='http://localhost', interval='30s'))
 
-    def test_register_script_and_no_interval(self):
+    def test_register_args_and_no_interval(self):
         with self.assertRaises(ValueError):
-            self.forbidden_consul.agent.check.register(
-                str(uuid.uuid4()),  '/bin/true')
+            self.consul.agent.check.register(
+                str(uuid.uuid4()),  args=['/bin/true'])
 
-    def test_register_script_and_ttl(self):
+    def test_register_args_and_ttl(self):
         with self.assertRaises(ValueError):
-            self.forbidden_consul.agent.check.register(
-                str(uuid.uuid4()), script='/bin/true', ttl=30)
+            self.consul.agent.check.register(
+                str(uuid.uuid4()), args=['/bin/true'], ttl='30s')
 
     def test_register_http_and_no_interval(self):
         with self.assertRaises(ValueError):
-            self.forbidden_consul.agent.check.register(
+            self.consul.agent.check.register(
                 str(uuid.uuid4()), http='http://localhost')
 
-    def test_register_script_and_http(self):
+    def test_register_args_and_http(self):
         with self.assertRaises(ValueError):
-            self.forbidden_consul.agent.check.register(
-                str(uuid.uuid4()), script='/bin/true', http='http://localhost')
+            self.consul.agent.check.register(
+                str(uuid.uuid4()), args=['/bin/true'], http='http://localhost')
 
     def test_register_forbidden(self):
         with self.assertRaises(consulate.Forbidden):
             self.forbidden_consul.agent.check.register(
-                str(uuid.uuid4()), '/bin/true', interval=30)
+                str(uuid.uuid4()), args=['/bin/true'], interval='30s')
 
 
 class TTLCheckTestCase(base.TestCase):
@@ -154,7 +155,7 @@ class TTLCheckTestCase(base.TestCase):
     def setUp(self):
         super(TTLCheckTestCase, self).setUp()
         name = str(uuid.uuid4())
-        self.assertTrue(self.consul.agent.check.register(name, ttl=30))
+        self.assertTrue(self.consul.agent.check.register(name, ttl='30s'))
         checks = self.consul.agent.checks()
         self.check_id = checks[name]['CheckID']
 
@@ -183,71 +184,112 @@ class TTLCheckTestCase(base.TestCase):
 class ServiceTestCase(base.TestCase):
 
     def test_register(self):
-        self.assertTrue(self.consul.agent.service.register(
-            str(uuid.uuid4()),
-            address='127.0.0.1',
-            port=80,
-            script='/bin/true',
-            interval=30,
-            tags=[str(uuid.uuid4())]))
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                check=agent.Check(
+                    name='test', args=['/bin/true'], interval='30s'),
+                tags=[str(uuid.uuid4())]))
+
+    def test_register_grpc(self):
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                check=agent.Check(
+                    name='test', grpc='https://grpc/status', interval='30s')))
 
     def test_register_http(self):
-        self.assertTrue(self.consul.agent.service.register(
-            str(uuid.uuid4()),
-            address='127.0.0.1',
-            port=80,
-            http='http://localhost',
-            interval=30,
-            tags=[str(uuid.uuid4())]))
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                check=agent.Check(
+                    name='test', http='http://localhost', interval='30s')))
+
+    def test_register_tcp(self):
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                check=agent.Check(
+                    name='test', tcp='localhost:80', interval='30s')))
 
     def test_register_ttl(self):
-        self.assertTrue(self.consul.agent.service.register(
-            str(uuid.uuid4()),
-            address='127.0.0.1',
-            port=80,
-            ttl=30,
-            tags=[str(uuid.uuid4())]))
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                check=agent.Check(name='test', ttl='30s')))
+
+    def test_register_multiple_checks(self):
+        self.assertTrue(
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                port=80,
+                checks=[
+                    agent.Check(
+                        name='test1', http='http://netloc',
+                        header={'User-Agent': 'unittest.TestCase'},
+                        interval='30s'),
+                    agent.Check(name='test2', ttl='30s')
+                ]))
 
     def test_register_forbidden(self):
         with self.assertRaises(consulate.Forbidden):
             self.forbidden_consul.agent.service.register(
                 str(uuid.uuid4()),
                 address='127.0.0.1',
-                port=80,
-                tags=[str(uuid.uuid4())])
+                port=80)
+
+    def test_register_invalid_check(self):
+        with self.assertRaises(TypeError):
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                check=str(uuid.uuid4()))
+
+    def test_register_invalid_checks(self):
+        with self.assertRaises(ValueError):
+            self.consul.agent.service.register(
+                str(uuid.uuid4()),
+                address='127.0.0.1',
+                checks=[str(uuid.uuid4())])
 
     def test_register_invalid_port(self):
-        with self.assertRaises(ValueError):
+        with self.assertRaises(TypeError):
             self.consul.agent.service.register(
                 str(uuid.uuid4()),
                 address='127.0.0.1',
-                port='80',
-                ttl=30,
-                tags=[str(uuid.uuid4())])
+                port='80')
 
     def test_register_invalid_tags(self):
-        with self.assertRaises(ValueError):
+        with self.assertRaises(TypeError):
             self.consul.agent.service.register(
                 str(uuid.uuid4()),
                 address='127.0.0.1',
-                ttl=30,
                 tags=str(uuid.uuid4()))
 
     def test_register_invalid_interval(self):
-        with self.assertRaises(ValueError):
+        with self.assertRaises(TypeError):
             self.consul.agent.service.register(
                 str(uuid.uuid4()),
                 address='127.0.0.1',
                 port=80,
-                http='http://localhost',
-                interval='30',
-                tags=[str(uuid.uuid4())])
+                check=agent.Check(
+                    name='test', http='http://localhost', interval=30))
 
     def test_register_invalid_ttl(self):
-        with self.assertRaises(ValueError):
+        with self.assertRaises(TypeError):
             self.consul.agent.service.register(
                 str(uuid.uuid4()),
                 address='127.0.0.1',
                 port=80,
-                ttl=-30,
-                tags=[str(uuid.uuid4())])
+                check=agent.Check(name='test', ttl=30))